+++ NEWS.md
@@ -3,7 +3,8 @@
 ## Version 0.2.5
   * replaced package RCurl with curl for downloading DWD Data
   * stopped importing package methods
-  * DWD changed layout of station description files: fixed read.DWDstations() 
+  * DWD changed layout of station description files: fixed read.DWDstations()
+  * DWD changes column order of climate data files: adapted documentation of read.DWDdata()
 
 
 ## Version 0.2.4